Fix parceling of seekbarProgress in SeekbarWidget
Changed writeToParcel to write seekbarProgress as an integer instead of a float.

+++ app/src/main/java/com/drdisagree/colorblendr/ui/widgets/SeekbarWidget.kt
@@ -275,7 +275,7 @@ class SeekbarWidget : RelativeLayout {
 
         override fun writeToParcel(dest: Parcel, flags: Int) {
             super.writeToParcel(dest, flags)
-            dest.writeFloat(seekbarProgress.toFloat())
+            dest.writeInt(seekbarProgress)
         }
 
         companion object CREATOR : Parcelable.Creator<SavedState> {
